change distance type to double instead of int

+++ lib/entry_view.dart
@@ -30,7 +30,7 @@ class _EntryView extends State<EntryView> {
   void _addEntry() async {
     DateTime? start;
     DateTime? end;
-    int? mileage;
+    double? mileage;
     List<String> tagList = <String>[];
 
     if (isTracking) {
@@ -69,7 +69,7 @@ class _EntryView extends State<EntryView> {
                   )),
               TextField(
                 onChanged: (value) {
-                  mileage = int.parse(value);
+                  mileage = double.parse(value);
                 },
                 decoration: const InputDecoration(hintText: "Enter mileage"),
               ),
@@ -112,7 +112,7 @@ class _EntryView extends State<EntryView> {
   void _editEntry(Entry entry, int index) async {
     DateTime? start = entry.start;
     DateTime? end = entry.end;
-    int? mileage = entry.mileage;
+    double? mileage = entry.mileage;
     List<String> tagList = entry.tagList;
 
     // Create TextEditingController for each TextField
@@ -162,7 +162,7 @@ class _EntryView extends State<EntryView> {
                 controller: mileageController, // Set the controller
                 onChanged: (value) {
                   // When the text changes, update the mileage
-                  mileage = int.parse(value);
+                  mileage = double.parse(value);
                 },
                 decoration: const InputDecoration(hintText: "Enter mileage"),
               ),
